+++ pac4j-core/src/test/java/org/pac4j/core/client/BaseClientTests.java
@@ -14,7 +14,7 @@ import static org.junit.Assert.*;
 
 /**
  * This class tests the {@link BaseClient} class.
- * 
+ *
  * @author Jerome Leleu
  * @since 1.4.0
  */
@@ -65,7 +65,7 @@ public final class BaseClientTests implements TestsConstants {
         final MockIndirectClient client = new MockIndirectClient(TYPE, RedirectAction.redirect(LOGIN_URL), (Credentials) null, new CommonProfile());
         client.setCallbackUrl(CALLBACK_URL);
         final MockWebContext context = MockWebContext.create();
-        context.setSessionAttribute(client.getName() + IndirectClient.ATTEMPTED_AUTHENTICATION_SUFFIX, "true");
+        context.getSessionStore().set(context, client.getName() + IndirectClient.ATTEMPTED_AUTHENTICATION_SUFFIX, "true");
         final HttpAction e = (HttpAction) TestsHelper.expectException(() -> client.redirect(context));
         assertEquals(401, e.getCode());
         assertEquals(401, context.getResponseStatus());
@@ -77,8 +77,7 @@ public final class BaseClientTests implements TestsConstants {
         client.setCallbackUrl(CALLBACK_URL);
         final MockWebContext context = MockWebContext.create();
         client.getCredentials(context);
-        assertEquals("true",
-                (String) context.getSessionAttribute(client.getName() + IndirectClient.ATTEMPTED_AUTHENTICATION_SUFFIX));
+        assertEquals("true", context.getSessionStore().get(context, client.getName() + IndirectClient.ATTEMPTED_AUTHENTICATION_SUFFIX));
     }
 
     @Test